근무 수정및 삭제 후 잔상 남는거 제거예정, 실제 근무시간 안보이는거 calendar.js
