+++ src/pages/Calendar/Calendar.js
@@ -8,12 +8,13 @@ import { v4 as uuidv4 } from "uuid";
 import ApiService from "../../ApiService";
 import CalendarInput from "./CalendarInput";
 import CalendarChart from "./CalendarChart";
+import CalendarDetail from "./CalendarDetail";
 
 const Calendar = () => {
   var groupSeq = Number(sessionStorage.getItem("group_seq"));
 
   var copyTodayWorkerList = [];
-  var copySelectedWorkerList = [];
+  // var copySelectedWorkerList = [];
   var date = new Date();
   var year = date.getFullYear();
   var month = ("0" + (date.getMonth() + 1)).slice(-2);
@@ -25,18 +26,18 @@ const Calendar = () => {
   const apiKey = "AIzaSyAHG8iIVB4i-q5o7KRjdvKcwVc67JzZEWc";
   const [selectedDate, setSelectedDate] = useState(today);
   const arrAddList = useRef([]);
-  const [thisDayListState, setThisDayListState] = useState([]);
-  const [arrAddListState, setArrAddListState] = useState([]);
+  // const [thisDayListState, setThisDayListState] = useState([]);
+  // const [arrAddListState, setArrAddListState] = useState([]);
   const [todayWorkerList, setTodayWorkerList] = useState([]);
   const selectedList = useRef([]);
   const workerSchedule = useRef([]);
-  const listConfirmation = useRef([]);
-  const listReject = useRef([]);
-  const rejectMemo = useRef();
+  // const listConfirmation = useRef([]);
+  // const listReject = useRef([]);
+  // const rejectMemo = useRef();
   const scheduleInfo = useRef([]);
   var copyScheduleInfo = [];
-  var copySchedule = [];
-  var copyModificationAllInfo = [];
+  // var copySchedule = [];
+  // var copyModificationAllInfo = [];
   const modificationAllInfo = useRef([]);
   const modificaionInfo = useRef([]);
   const workerInfo = useRef([]);
@@ -48,16 +49,16 @@ const Calendar = () => {
     getSchedule(groupSeq);
   }, []);
 
-  const confirmModification = (e) => {
-    ApiService.confirmModification(e).then(() => {
-      getSchedule(groupSeq);
-    });
-  };
-  const rejectModification = (e) => {
-    ApiService.rejectModification(e).then(() => {
-      getSchedule(groupSeq);
-    });
-  };
+  // const confirmModification = (e) => {
+  //   ApiService.confirmModification(e).then(() => {
+  //     getSchedule(groupSeq);
+  //   });
+  // };
+  // const rejectModification = (e) => {
+  //   ApiService.rejectModification(e).then(() => {
+  //     getSchedule(groupSeq);
+  //   });
+  // };
 
   const getModification = (groupSeq) => {
     modificaionInfo.current = [];
@@ -135,16 +136,16 @@ const Calendar = () => {
       });
   };
 
-  const updateSchedul = (e) => {
-    ApiService.updateSchedul(e)
-      .then(() => {
-        console.log("업데이트성공");
-        getSchedule(groupSeq);
-      })
-      .catch((err) => {
-        alert(err);
-      });
-  };
+  // const updateSchedul = (e) => {
+  //   ApiService.updateSchedul(e)
+  //     .then(() => {
+  //       console.log("업데이트성공");
+  //       getSchedule(groupSeq);
+  //     })
+  //     .catch((err) => {
+  //       alert(err);
+  //     });
+  // };
 
   workerSchedule.current.length > 0
     ? (copyTodayWorkerList = [...workerSchedule.current])
@@ -154,11 +155,12 @@ const Calendar = () => {
     copyTodayWorkerList.push(modificaionInfo.current[i]);
   }
 
-  copyModificationAllInfo = [...modificationAllInfo.current];
+  // copyModificationAllInfo = [...modificationAllInfo.current];
 
-  copySelectedWorkerList = [...selectedList.current];
+  // copySelectedWorkerList = [...selectedList.current];
 
   copyScheduleInfo = [scheduleInfo.current];
+  console.log("카피tmzpw", copyScheduleInfo);
 
   const clickDate = (e) => {
     selectedList.current = [];
@@ -193,52 +195,52 @@ const Calendar = () => {
     setSelectedDate(String(e));
   };
 
-  const planModification = () => {
-    var result = copySelectedWorkerList.map((item, index) => {
-      return (
-        <tr key={`${item.mem_name}${index}`}>
-          <select
-            id="select"
-            onChange={(e) => {
-              copySelectedWorkerList[index].mem_name = e.target.value;
-              copySelectedWorkerList[index].group_seq = groupSeq;
-            }}
-            defaultValue={item.mem_name}
-          >
-            {workerListRendering()}
-          </select>
-          <input
-            id="inputTime"
-            type="time"
-            defaultValue={item.att_sche_start_time}
-            onChange={(e) => {
-              copySelectedWorkerList[index].att_sche_start_ime = e.target.value;
-            }}
-          />
-          -
-          <input
-            id="inputTime"
-            type="time"
-            defaultValue={item.att_sche_end_time}
-            onChange={(e) => {
-              copySelectedWorkerList[index].att_sche_end_time = e.target.value;
-            }}
-          />
-          <button
-            id="delete"
-            onClick={() => {
-              selectedList.current.splice(index, 1);
-              setThisDayListState([selectedList.current]);
-            }}
-          >
-            삭제
-          </button>
-        </tr>
-      );
-    });
+  // const planModification = () => {
+  //   var result = copySelectedWorkerList.map((item, index) => {
+  //     return (
+  //       <tr key={`${item.mem_name}${index}`}>
+  //         <select
+  //           id="select"
+  //           onChange={(e) => {
+  //             copySelectedWorkerList[index].mem_name = e.target.value;
+  //             copySelectedWorkerList[index].group_seq = groupSeq;
+  //           }}
+  //           defaultValue={item.mem_name}
+  //         >
+  //           {workerListRendering()}
+  //         </select>
+  //         <input
+  //           id="inputTime"
+  //           type="time"
+  //           defaultValue={item.att_sche_start_time}
+  //           onChange={(e) => {
+  //             copySelectedWorkerList[index].att_sche_start_ime = e.target.value;
+  //           }}
+  //         />
+  //         -
+  //         <input
+  //           id="inputTime"
+  //           type="time"
+  //           defaultValue={item.att_sche_end_time}
+  //           onChange={(e) => {
+  //             copySelectedWorkerList[index].att_sche_end_time = e.target.value;
+  //           }}
+  //         />
+  //         <button
+  //           id="delete"
+  //           onClick={() => {
+  //             selectedList.current.splice(index, 1);
+  //             setThisDayListState([selectedList.current]);
+  //           }}
+  //         >
+  //           삭제
+  //         </button>
+  //       </tr>
+  //     );
+  //   });
 
-    return result;
-  };
+  //   return result;
+  // };
 
   const workerListRendering = () => {
     var result = workerInfo.current.map((item, index) => {
@@ -305,261 +307,261 @@ const Calendar = () => {
     );
   };
 
-  const addModification = () => {
-    let result = arrAddList.current.map((item, index) => {
-      return (
-        <tr key={uuidv4()}>
-          <select
-            id="select"
-            defaultValue={item.mem_name}
-            onChange={(e) => {
-              console.log(e.target.value);
-              for (var i = 0; i < workerInfo.current.length; i++) {
-                if (e.target.value === workerInfo.current[i].mem_name) {
-                  arrAddList.current[index].mem_name = e.target.value;
-                  arrAddList.current[index].mem_id =
-                    workerInfo.current[i].mem_id;
-                }
-              }
-            }}
-          >
-            {workerListRendering()}
-          </select>
-          <input
-            type="time"
-            defaultValue={item.att_sche_start_time}
-            onChange={(e) => {
-              arrAddList.current[index].att_sche_start_time = e.target.value;
-            }}
-          />
-          ~
-          <input
-            defaultValue={item.att_sche_end_time}
-            type="time"
-            onChange={(e) => {
-              arrAddList.current[index].att_sche_end_time = e.target.value;
-            }}
-          />
-          <button
-            id="delete"
-            onClick={() => {
-              arrAddList.current.splice(index, 1);
-              setArrAddListState([arrAddList.current]);
-            }}
-          >
-            삭제
-          </button>
-        </tr>
-      );
-    });
-    return result;
-  };
+  // const addModification = () => {
+  //   let result = arrAddList.current.map((item, index) => {
+  //     return (
+  //       <tr key={uuidv4()}>
+  //         <select
+  //           id="select"
+  //           defaultValue={item.mem_name}
+  //           onChange={(e) => {
+  //             console.log(e.target.value);
+  //             for (var i = 0; i < workerInfo.current.length; i++) {
+  //               if (e.target.value === workerInfo.current[i].mem_name) {
+  //                 arrAddList.current[index].mem_name = e.target.value;
+  //                 arrAddList.current[index].mem_id =
+  //                   workerInfo.current[i].mem_id;
+  //               }
+  //             }
+  //           }}
+  //         >
+  //           {workerListRendering()}
+  //         </select>
+  //         <input
+  //           type="time"
+  //           defaultValue={item.att_sche_start_time}
+  //           onChange={(e) => {
+  //             arrAddList.current[index].att_sche_start_time = e.target.value;
+  //           }}
+  //         />
+  //         ~
+  //         <input
+  //           defaultValue={item.att_sche_end_time}
+  //           type="time"
+  //           onChange={(e) => {
+  //             arrAddList.current[index].att_sche_end_time = e.target.value;
+  //           }}
+  //         />
+  //         <button
+  //           id="delete"
+  //           onClick={() => {
+  //             arrAddList.current.splice(index, 1);
+  //             setArrAddListState([arrAddList.current]);
+  //           }}
+  //         >
+  //           삭제
+  //         </button>
+  //       </tr>
+  //     );
+  //   });
+  //   return result;
+  // };
 
-  const pushArrAddList = () => {
-    arrAddList.current.push({
-      mem_name: "",
-      att_sche_start_time: "",
-      att_sche_end_time: "",
-      mem_id: "",
-      group_seq: groupSeq,
-    });
-  };
+  // const pushArrAddList = () => {
+  //   arrAddList.current.push({
+  //     mem_name: "",
+  //     att_sche_start_time: "",
+  //     att_sche_end_time: "",
+  //     mem_id: "",
+  //     group_seq: groupSeq,
+  //   });
+  // };
 
-  const submitModification = () => {
-    var updateInfo = copySelectedWorkerList.concat(arrAddList.current);
+  // const submitModification = () => {
+  //   var updateInfo = copySelectedWorkerList.concat(arrAddList.current);
 
-    for (var i = 0; i < updateInfo.length; i++) {
-      updateInfo[i]["att_date"] = selectedDate;
-    }
+  //   for (var i = 0; i < updateInfo.length; i++) {
+  //     updateInfo[i]["att_date"] = selectedDate;
+  //   }
 
-    arrAddList.current = [];
+  //   arrAddList.current = [];
 
-    if (updateInfo.length > 0) {
-      updateSchedul(updateInfo);
-    } else {
-      updateInfo = [
-        {
-          att_date: selectedDate,
-          group_seq: groupSeq,
-          mem_id: "010",
-          att_sche_start_time: "10:00",
-          att_sche_end_time: "10:00",
-        },
-      ];
+  //   if (updateInfo.length > 0) {
+  //     updateSchedul(updateInfo);
+  //   } else {
+  //     updateInfo = [
+  //       {
+  //         att_date: selectedDate,
+  //         group_seq: groupSeq,
+  //         mem_id: "010",
+  //         att_sche_start_time: "10:00",
+  //         att_sche_end_time: "10:00",
+  //       },
+  //     ];
 
-      updateSchedul(updateInfo);
-    }
-    console.log("최종배열", updateInfo);
-  };
+  //     updateSchedul(updateInfo);
+  //   }
+  //   console.log("최종배열", updateInfo);
+  // };
 
-  const clickDetail = () => {
-    for (var i = 0; i < copySchedule.length; i++) {
-      if (copySchedule[i].date === selectedDate.replace(/-/gi, "")) {
-        return copySchedule[i].mem_name;
-      }
-    }
-  };
+  // const clickDetail = () => {
+  //   for (var i = 0; i < copySchedule.length; i++) {
+  //     if (copySchedule[i].date === selectedDate.replace(/-/gi, "")) {
+  //       return copySchedule[i].mem_name;
+  //     }
+  //   }
+  // };
 
-  const planSchedule = () => {
-    var result = copySelectedWorkerList.map((item, index) => {
-      return (
-        <>
-          <tr key={uuidv4}>
-            <th> 계획</th>
-            <th>{item.mem_name}</th>
-            <th>{item.att_sche_start_time.substring(0, 5)}</th>
-            <th>{item.att_sche_end_time.substring(0, 5)}</th>
-          </tr>
+  // const planSchedule = () => {
+  //   var result = copySelectedWorkerList.map((item, index) => {
+  //     return (
+  //       <>
+  //         <tr key={uuidv4}>
+  //           <th> 계획</th>
+  //           <th>{item.mem_name}</th>
+  //           <th>{item.att_sche_start_time.substring(0, 5)}</th>
+  //           <th>{item.att_sche_end_time.substring(0, 5)}</th>
+  //         </tr>
 
-          <tr key={uuidv4}>
-            <th> 실제 </th>
-            <th>{item.mem_name}</th> <th>{item.att_real_start_time}</th>
-            <th>{item.att_real_end_time}</th>
-          </tr>
-          <br></br>
-        </>
-      );
-    });
+  //         <tr key={uuidv4}>
+  //           <th> 실제 </th>
+  //           <th>{item.mem_name}</th> <th>{item.att_real_start_time}</th>
+  //           <th>{item.att_real_end_time}</th>
+  //         </tr>
+  //         <br></br>
+  //       </>
+  //     );
+  //   });
 
-    return (
-      <>
-        <table width="100%">
-          <tr>
-            <th>
-              <h2>구분</h2>
-            </th>
-            <th>
-              <h2>이름</h2>
-            </th>
-            <th>
-              <h2>출근</h2>
-            </th>
-            <th>
-              <h2>퇴근</h2>
-            </th>
-          </tr>
+  //   return (
+  //     <>
+  //       <table width="100%">
+  //         <tr>
+  //           <th>
+  //             <h2>구분</h2>
+  //           </th>
+  //           <th>
+  //             <h2>이름</h2>
+  //           </th>
+  //           <th>
+  //             <h2>출근</h2>
+  //           </th>
+  //           <th>
+  //             <h2>퇴근</h2>
+  //           </th>
+  //         </tr>
 
-          {result}
-        </table>
-      </>
-    );
-  };
+  //         {result}
+  //       </table>
+  //     </>
+  //   );
+  // };
 
-  const showModificaion = () => {
-    console.log("오늘", today);
-    console.log("선택", selectedDate);
-    console.log("결과", selectedDate < today);
-    if (selectedDate < today) return planSchedule();
-    else {
-      return planModification();
-    }
-  };
+  // const showModificaion = () => {
+  //   console.log("오늘", today);
+  //   console.log("선택", selectedDate);
+  //   console.log("결과", selectedDate < today);
+  //   if (selectedDate < today) return planSchedule();
+  //   else {
+  //     return planModification();
+  //   }
+  // };
 
-  const addButton = () => {
-    if (selectedDate < today) {
-      return;
-    } else {
-      return (
-        <button
-          id="registerButton"
-          onClick={() => {
-            pushArrAddList();
-            setArrAddListState([arrAddList.current]);
-          }}
-        >
-          추가
-        </button>
-      );
-    }
-  };
+  // const addButton = () => {
+  //   if (selectedDate < today) {
+  //     return;
+  //   } else {
+  //     return (
+  //       <button
+  //         id="registerButton"
+  //         onClick={() => {
+  //           pushArrAddList();
+  //           setArrAddListState([arrAddList.current]);
+  //         }}
+  //       >
+  //         추가
+  //       </button>
+  //     );
+  //   }
+  // };
 
-  const registerButton = () => {
-    if (selectedDate < today) {
-      return;
-    } else {
-      return (
-        <input
-          id="registerButton"
-          type="submit"
-          name="등록"
-          value="등록"
-          onClick={submitModification}
-        />
-      );
-    }
-  };
+  // const registerButton = () => {
+  //   if (selectedDate < today) {
+  //     return;
+  //   } else {
+  //     return (
+  //       <input
+  //         id="registerButton"
+  //         type="submit"
+  //         name="등록"
+  //         value="등록"
+  //         onClick={submitModification}
+  //       />
+  //     );
+  //   }
+  // };
 
-  const modificationAnser = () => {
-    listReject.current = [];
-    listConfirmation.current = [];
-    var result = copyModificationAllInfo.map((item, index) => {
-      if (selectedDate === copyModificationAllInfo[index].ch_date) {
-        return (
-          <table width="100%">
-            <br />
-            <br />
-            <h2 style={{ color: "red" }}>근무변경요청</h2>
-            <tr key={uuidv4} width="100%">
-              <th>
-                {" "}
-                {item.mem_name} :{item.ch_start_time.substring(0, 5)}-
-                {item.ch_end_time.substring(0, 5)}
-              </th>
-            </tr>
-            <tr key={uuidv4} width="100%">
-              <th>
-                <input
-                  id="inputText"
-                  width="100%"
-                  type="textArea"
-                  placeholder="거절시 사유를 적어주세요."
-                  onChange={(e) => {
-                    rejectMemo.current = e.target.value;
-                    console.log(rejectMemo.current);
-                  }}
-                ></input>
-              </th>
-            </tr>
-            <tr>
-              <button
-                id="registerButton"
-                onClick={() => {
-                  listConfirmation.current.push({
-                    ch_seq: item.ch_seq,
-                    att_seq: item.att_seq,
-                    ch_approve: "Y",
-                    mem_id: item.mem_id,
-                    ch_start_time: item.ch_start_time,
-                    ch_end_time: item.ch_end_time,
-                    ch_date: item.ch_date,
-                    group_seq: item.group_seq,
-                  });
+  // const modificationAnser = () => {
+  //   listReject.current = [];
+  //   listConfirmation.current = [];
+  //   var result = copyModificationAllInfo.map((item, index) => {
+  //     if (selectedDate === copyModificationAllInfo[index].ch_date) {
+  //       return (
+  //         <table width="100%">
+  //           <br />
+  //           <br />
+  //           <h2 style={{ color: "red" }}>근무변경요청</h2>
+  //           <tr key={uuidv4} width="100%">
+  //             <th>
+  //               {" "}
+  //               {item.mem_name} :{item.ch_start_time.substring(0, 5)}-
+  //               {item.ch_end_time.substring(0, 5)}
+  //             </th>
+  //           </tr>
+  //           <tr key={uuidv4} width="100%">
+  //             <th>
+  //               <input
+  //                 id="inputText"
+  //                 width="100%"
+  //                 type="textArea"
+  //                 placeholder="거절시 사유를 적어주세요."
+  //                 onChange={(e) => {
+  //                   rejectMemo.current = e.target.value;
+  //                   console.log(rejectMemo.current);
+  //                 }}
+  //               ></input>
+  //             </th>
+  //           </tr>
+  //           <tr>
+  //             <button
+  //               id="registerButton"
+  //               onClick={() => {
+  //                 listConfirmation.current.push({
+  //                   ch_seq: item.ch_seq,
+  //                   att_seq: item.att_seq,
+  //                   ch_approve: "Y",
+  //                   mem_id: item.mem_id,
+  //                   ch_start_time: item.ch_start_time,
+  //                   ch_end_time: item.ch_end_time,
+  //                   ch_date: item.ch_date,
+  //                   group_seq: item.group_seq,
+  //                 });
 
-                  confirmModification(listConfirmation.current);
-                }}
-              >
-                승락
-              </button>
-              <button
-                id="registerButton"
-                onClick={() => {
-                  listReject.current.push({
-                    ch_seq: item.ch_seq,
-                    ch_approve: "N",
-                    ch_reject_memo: rejectMemo.current,
-                  });
-                  rejectModification(listReject.current);
-                }}
-              >
-                거절
-              </button>
-            </tr>
-          </table>
-        );
-      }
-    });
-    return result;
-  };
+  //                 confirmModification(listConfirmation.current);
+  //               }}
+  //             >
+  //               승락
+  //             </button>
+  //             <button
+  //               id="registerButton"
+  //               onClick={() => {
+  //                 listReject.current.push({
+  //                   ch_seq: item.ch_seq,
+  //                   ch_approve: "N",
+  //                   ch_reject_memo: rejectMemo.current,
+  //                 });
+  //                 rejectModification(listReject.current);
+  //               }}
+  //             >
+  //               거절
+  //             </button>
+  //           </tr>
+  //         </table>
+  //       );
+  //     }
+  //   });
+  //   return result;
+  // };
 
   const showModal = () => {
     setModalOpen(true);
@@ -628,8 +630,21 @@ const Calendar = () => {
           }}
         />
       </div>
+      <CalendarDetail
+        today={today}
+        modificaionInfo={modificaionInfo.current}
+        selectedList={selectedList.current}
+        getSchedule={getSchedule}
+        groupSeq={groupSeq}
+        modificationAllInfo={modificationAllInfo.current}
+        arrAddList={arrAddList.current}
+        selectedDate={selectedDate}
+        workerListRendering={workerListRendering}
+        workerInfo={workerInfo.current}
+        setSelectedDate={setSelectedDate}
+      />
 
-      <div className="calendarDetailContainer">
+      {/* <div className="calendarDetailContainer">
         <div className="modification">
           <table style={{ width: "100%" }}>
             <tr>
@@ -637,7 +652,7 @@ const Calendar = () => {
             </tr>
             <tr align="left">
               <h3>
-                {/* {mem_name} */}
+               
                 {clickDetail()}
               </h3>
             </tr>
@@ -651,8 +666,8 @@ const Calendar = () => {
           <tr></tr>
         </div>
         <div className="special">{modificationAnser()}</div>
-        {/* <button onClick={showChart}>직원차트</button> */}
-      </div>
+       
+      </div> */}
     </div>
   );
 };
